requirements.txt
# Core: Tkinter is bundled with Python on Windows

# Optional notifications
plyer>=2.1.0; platform_system == "Windows"

# Optional system tray support
pystray>=0.19.5
Pillow>=10.0.0
